handle_host_links task; fetches and parses successfully
