+++ build.gradle.kts
@@ -1,15 +1,3 @@
-import com.google.gson.Gson
-import com.google.gson.JsonObject
-
-buildscript {
-    repositories {
-        google()
-    }
-    dependencies {
-        classpath("com.google.code.gson:gson:2.10.1")
-    }
-}
-
 plugins {
     // trick: for the same plugin versions in all sub-modules
     id("com.android.application").version(Versions.androidPlugin).apply(false)
@@ -26,36 +14,9 @@ tasks.register("clean", Delete::class) {
     delete(rootProject.buildDir)
 }
 
-// TODO separating this logic into its own gradle.kts file breaks the gson dependency and i don't know why
-// TODO rename to handle_host_links
-tasks.register("temp") {
-    val tempPath = "${rootProject.buildDir}/hosts.txt"
-    val hosts = fetchHosts(tempPath)
-    androidHosts(hosts)
-    iosHosts(hosts)
-    cleanup(tempPath)
+tasks.register("handle_host_links") {
+    val buildPath = "$projectDir/buildSrc/build/tmp/hosts.txt"
+    val androidPath = "$projectDir/androidApp/src/main"
+    val iosPath = "" // TODO
+    handleHostLinks(buildPath, androidPath, iosPath)
 }
-
-fun fetchHosts(path: String): List<String> {
-    val file = file(path).also { it.createNewFile() }
-    val count = 50
-    val cmd = listOf("curl", "-X", "GET",
-        "-H", "Authorization: Bearer ${System.getenv("INSTANCES_SOCIAL_API_KEY")}",
-        "https://instances.social/api/1.0/instances/list?count=$count&sort_by=users&sort_order=desc",
-        "-o", path)
-    ProcessBuilder(cmd).start().waitFor()
-
-    val json = Gson().fromJson(file.readText(), JsonObject::class.java)
-    return json.get("instances").asJsonArray.map { it.asJsonObject["name"].asString }.toList()
-}
-
-fun androidHosts(hosts: List<String>) {
-    println("Creating host handling for Android...")
-    val androidDir = "$projectDir/androidApp/src/main"
-}
-
-fun iosHosts(hosts: List<String>) {
-    println("Creating host handling for iOS...")
-}
-
-fun cleanup(path: String) = file(path).delete()
