+++ buildSrc/build.gradle.kts
@@ -7,3 +7,7 @@ repositories {
 plugins {
     `kotlin-dsl`
 }
+
+dependencies {
+    implementation("com.google.code.gson:gson:2.10.1")
+}
